+++ astc/astc_pick_best_endpoint_format.cpp
@@ -178,7 +178,7 @@ static void compute_color_error_for_every_integer_count_and_quantization_level(i
 		float3 prd = ep1.xyz - float3(cf, cf, cf);
 		float3 pdif = prd - ep0.xyz;
 		// estimate of color-component spread in low endpoint color
-		float df = MAX(MAX(fabs(pdif.x), fabs(pdif.y)), fabs(pdif.z));
+		float df = MAX(MAX(fabsf(pdif.x), fabsf(pdif.y)), fabsf(pdif.z));
 
 		int b = (int)bf;
 		int c = (int)cf;
